Move search-related views to namespace, move search details to view
