+++ PalCalc.UI/View/Inspector/Search/ContainerGridView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="PalCalc.UI.View.Inspector.ContainerGridView"
+﻿<UserControl x:Class="PalCalc.UI.View.Inspector.Search.ContainerGridView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
