+++ PalCalc.UI/View/Inspector/Search/ContainerGridView.xaml.cs
@@ -1,0 +1,29 @@
+﻿using PalCalc.UI.ViewModel.Inspector.Search;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace PalCalc.UI.View.Inspector.Search
+{
+    /// <summary>
+    /// Interaction logic for ContainerGridView.xaml
+    /// </summary>
+    public partial class ContainerGridView : UserControl
+    {
+        public ContainerGridView()
+        {
+            InitializeComponent();
+        }
+    }
+}
