+++ PalCalc.UI/View/Inspector/Search/CustomPalEditorView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="PalCalc.UI.View.Inspector.CustomPalEditorView"
+﻿<UserControl x:Class="PalCalc.UI.View.Inspector.Search.CustomPalEditorView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
